+++ ui/views/controls/textfield/textfield_model.cc
@@ -69,9 +69,10 @@ class Edit {
     std::reverse(insertion_texts.begin(), insertion_texts.end());
     auto insertion_text_starts = old_text_starts_;
     std::reverse(insertion_text_starts.begin(), insertion_text_starts.end());
-    model->ModifyText({{new_text_start_, new_text_end()}}, insertion_texts,
-                      insertion_text_starts, old_primary_selection_,
-                      old_secondary_selections_);
+    model->ModifyText({{static_cast<uint32_t>(new_text_start_),
+                        static_cast<uint32_t>(new_text_end())}},
+                      insertion_texts, insertion_text_starts,
+                      old_primary_selection_, old_secondary_selections_);
   }
 
   // Apply the change of this edit to the |model|.
@@ -82,7 +83,9 @@ class Edit {
                              old_text_starts_[i] + old_texts_[i].length());
     }
     model->ModifyText(deletions, {new_text_}, {new_text_start_},
-                      {new_cursor_pos_, new_cursor_pos_}, {});
+                      {static_cast<uint32_t>(new_cursor_pos_),
+                       static_cast<uint32_t>(new_cursor_pos_)},
+                      {});
   }
 
   // Try to merge the |edit| into this edit and returns true on success. The
